requirements.txt
fastmcp>=2.12.0
uvicorn>=0.35.0
python-dotenv==1.1.1
requests==2.32.5
fastapi==0.116.1